+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 2.7.3
+
+ - Add a "fill wand" feature which will add all known spells to a wand when creating.
+
 ## 2.7.2
  
  - Added effects to fill and blink.